Remove explicit /usr/bin from PATH
Not exactly sure what's going on here, but having `/usr/bin` in my $PATH
makes vim kinda unhappy when trying to run `!ruby commands`. This is
because the path order is different in vim... and it only happens on
macOS because there are system-provided binaries and scripts like
`/usr/bin/ruby` and `/usr/bin/rails` for some reason. None of this is
good.

So yeah, this is my way around all that.