Add Vim grep to quickfix window function
When you run this function and you have text visually selected, it
searches your current project context for the selected text.

If you do not have text visually selected, the function prompts you for
a regular expression.

Results are posted to a quickfix window, which allows you to select and
navigate directly to the results.