You don't need `kcr`
kcr is incredibly good and cool, but it turns out all I *really* need
from it is the `connect` command with a few modifications. So, with this
commit, that's what I've done.

In making this change, I've moved all of the scripts I used with `kcr`
to a new location, and I've removed any dependencies they had on `jq`
and `kcr`. This means that bootstrapping my next computer just got a lot
simpler.